Add enablePreview option, move options to context.
