+++ src/components/Items/Items.tsx
@@ -1,6 +1,6 @@
 import React, { useRef } from "react";
 import { CollectionItems, Collection, Manifest } from "@iiif/presentation-3";
-import { FetchCredentials, SwiperBreakpoints } from "../../../types/types";
+import { SwiperBreakpoints } from "../../types/types";
 import Item from "./Item";
 import { ItemsStyled } from "./Items.styled";
 import { Navigation, A11y } from "swiper";
@@ -8,7 +8,6 @@ import { Swiper, SwiperSlide } from "swiper/react";
 
 interface ItemsProps {
   breakpoints?: SwiperBreakpoints;
-  credentials: FetchCredentials;
   instance: number;
   items: CollectionItems[];
 }
@@ -43,7 +42,6 @@ const defaultBreakpoints = {
 
 const Items: React.FC<ItemsProps> = ({
   breakpoints = defaultBreakpoints,
-  credentials,
   instance,
   items,
 }) => {
@@ -72,11 +70,7 @@ const Items: React.FC<ItemsProps> = ({
             data-index={index}
             data-type={item?.type.toLowerCase()}
           >
-            <Item
-              credentials={credentials}
-              index={index}
-              item={item as Collection | Manifest}
-            />
+            <Item index={index} item={item as Collection | Manifest} />
           </SwiperSlide>
         ))}
       </Swiper>
